Renamed NotificationSensorListenerService NotificationSensorManager
